+++ .storybook/main.ts
@@ -5,9 +5,10 @@ const config: StorybookConfig = {
   stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
   addons: [
     "@storybook/addon-links",
-    "@chromatic-com/storybook",
     "@storybook/addon-designs",
-    "@storybook/addon-docs"
+    "@storybook/addon-docs",
+    "@storybook/addon-a11y",
+    "storybook-addon-pseudo-states"
   ],
   framework: {
     name: "@storybook/react-vite",