+++ src/components/inputs/DatePicker/DatePicker.stories.tsx
@@ -4,7 +4,7 @@ import { Box } from "@mui/material";
 import { useState } from "react";
 import dayjs, { Dayjs } from "dayjs";
 
-const meta = {
+const meta: Meta<typeof DatePicker> = {
   title: "Inputs/DatePicker",
   component: DatePicker,
   tags: ["autodocs"],
@@ -41,7 +41,7 @@ const [date, setDate] = useState(dayjs());
     },
   },
   decorators: [(Story) => <Box sx={{ width: "100%", mx: "auto" }}>{Story()}</Box>],
-} satisfies Meta<typeof DatePicker>;
+};
 
 export default meta;
 type Story = StoryObj<typeof DatePicker>;